Remove us-east-1 bias from templates/docs
Created a deploy bucket in each region and updated the deploy scripts to also deploy to each region. This enables the CFN templates to work in any region. Previously would get an error stating the bucket with the Lambda code was in another region. Also updated docs to remove us-east-1 from the quick-create links.