[s3] base compression on Content-Encoding unless forced
In the past we've tried to support processing gzip-compressed objects based on content-type or filename extension. This was bound to run into problems, and it has: we currently treat `application/octet-stream` as gzip-compressed, but that's wrong. We should be looking at the `Content-Encoding` metadata on the s3 object to determine if it's gzip compressed. The AWS SDK (via the default HTTP transport) does this by default, actually, so we don't even need to do it ourselves if the objects are set with the correct encoding.

Unfortunately, plenty of systems don't set the encoding correctly, which leaves us attempting to make guesses about the compression type. Repeatedly guessing compression type requires multiple S3 reads and is not really efficient, so I'd rather follow the correct behavior (respecting Content-Encoding) and allow an override if we want to try and decompress objects that don't set the encoding header. This will break users relying on the previous `*.gz` or `application/octet-stream` logic, and we'll need to direct them to use the new `ForceGunzip` option.