[postgresql] Fix sampling in postgresql handler
The postgres handler sends events with `SendPresampled`, assuming the postgres parser is doing the sampling. I think this was some copy-pasta from the mysql implementation. THe postgresql parser doesn't currently do sampling, but the mysql parser does.

Calling `Send()` will use the sample rate initialized in libhoney, which is set in the SAMPLE_RATE env var read in by InitHoneycombFromEnvVars.