feat: rename fields for alb ingest
We use the rename-fields flag for honeytail and really like it. Adding
the same support here for alb.

Should I add this for the other tools too? ALB is the only one I am
using and can test.

Is there a way I can write a unit test for this?