publish every main build to S3 (#161)
## Which problem is this PR solving?

I'd like to be able to test main-line builds of this repo. Right now,
afaict this requires pulling down the code and manually
building/publishing the resulting zip to S3 somewhere.

## Short description of the changes

Instead, change CI to push every build to S3. It appears that
`publish_aws.sh` already has support for non-tagged builds in the form
of
[GIT_VERSION](https://github.com/honeycombio/agentless-integrations-for-aws/blob/main/publish_aws.sh#L10-L11),
so all that needs to be done is to run that script on any commit to
main.