get cloudfront time parsing working
the date & time need to come in as one field so that we can point the
httime parser at it. Additionally, we need to know the format of that
field. When using a specific `parser_type` that's easy to know, so
hard-code the appropriate field name & field format for the parsers we
control.