Add CloudFront regex parser (#166)
Add a basic regex for CloudFront logs.

This only supports the "everything" CF access log format, not anything
that's been configured with specific fields. That would require parsing
the header line at the top of each file with the exact fields in the
log, which is ... harder.

Unfortunately, there is still no testing for these regexps, so ✋ I
hereby certify I tested this regex against an actual CloudFront log
record.