[s3-handler] Adds support for sample rates based on prefixes
Post discussion on slack, this adds support for different sample rates
for a single bucket based on prefixes. We pass in this information via
an ENV var that is formatted as a JSON object. This is ugly, but the
rationale is that most users aren't going to be interacting with this
directly, but rather through the terraform module, and we can provide
a convenient interface for it through that.

Some decisions I felt are worth calling out:
- If the env var is non-empty and is invalid JSON, we error out. I
figured this is what _I_ would want in that situation.
- First rule that matches wins. I considered doing a
longest-match-wins scheme, but uhhhhh _nah_
- The format might look a little strange: one might expect it to be an
object like `{key: samplerate}`. I opted against this, because I
imagined we might want to add capabilities like having ALB-level
rules, or response status code level rules, etc. In theory support for
those is just adding new fields to the existing sampleRateRule struct.