feat: [s3-handler] Adds support for sample rates based on prefixes (#205)
Post discussion on slack, this adds support for different sample rates
for a single bucket based on prefixes. We pass in this information via
an ENV var that is formatted as a JSON object. This is ugly, but the
rationale is that most users aren't going to be interacting with this
directly, but rather through the terraform module, and we can provide a
convenient interface for it through that.

Some decisions I felt are worth calling out:
- If the env var is non-empty and is invalid JSON, we error out. I
figured this is what _I_ would want in that situation.
- First rule that matches wins. I considered doing a longest-match-wins
scheme, but uhhhhh _nah_
- The format might look a little strange: one might expect it to be an
object like `{key: samplerate}`. I opted against this, because I
imagined we might want to add capabilities like having ALB-level rules,
or response status code level rules, etc. In theory support for those is
just adding new fields to the existing sampleRateRule struct.
- If no rule applies, we fall back to the SAMPLE_RATE env var


## How I tested this

I configured the lambda like this:

```
    SAMPLE_RATE         = 2
    SAMPLE_RATE_RULES = jsonencode([
      {
        Prefix: "sampled-3/",
        SampleRate: 3
      },
      {
        Prefix: "unsampled/",
        SampleRate: 1,
      }
    ])
```

I then created 3 ALBs:
* `unsampled`
* `sampled-3`
* `not-in-rules`

Each ALB was configured to send their logs to a prefix in the S3 bucket
with the same name. I sent the data to my personal team in honeycomb and
queried the sample rate


https://ui.honeycomb.io/nlincoln-test-in-prod/environments/aws-lambda-testing/datasets/aws-elb-access/usage/result/p1u4yEh1Dgu
(nobody but me has access to that, ping me if you want to get added)

<img width="1147" alt="image"
src="https://github.com/honeycombio/agentless-integrations-for-aws/assets/2380692/f12cae04-48a8-4437-87fa-97ea4e13d896">


(Ignore sampled-2. It works but doesn't show off that the thing works)