chore: upgrade go1.x to provided.al2 runtime and use arm64 arch (#220)
## Which problem is this PR solving?

#218 

## Short description of the changes

Necessary changes to the build pipeline to support upgrading to the
`provided.al2` runtime, using [this AWS
blog](https://aws.amazon.com/blogs/compute/migrating-aws-lambda-functions-from-the-go1-x-runtime-to-the-custom-runtime-on-amazon-linux-2/)
as guidance. AWS is deprecating `go1.x` on December 31st. This approach
continues versioning and releasing these binaries as a unit but uploads
each individual zip file to S3 and the GitHub release rather than one
big `ingest-handlers.zip`.

Also snuck in a change from amd64 to arm64 to hopefully take advantage
of the cost + performance savings Amazon mentions as a motivator for
this change.

I tested the build script in isolation, the publish aws script in
isolation, and tested a binary e2e to validate events still get to
Honeycomb with the new runtime.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1205446566084103