feat: Create sns-kfh-transform[er] (#227)
## Which problem is this PR solving?

We want to use `aws_db_event_subscription` to get `db-instance` and
`db-snapshot` events into honeycomb. Since the data being sent is an
[events.SNSEntity](https://github.com/aws/aws-lambda-go/blob/main/events/sns.go#L20-L32),
and that struct's Message field is a json string, we need to transform
it to a struct so that it can be properly unpacked by honeycomb at the
receiving end.

## Short description of the changes
Part of work being done at
https://app.asana.com/0/1204553808824829/1206131228284410.

For examples of this data coming through (on a manually-deployed lambda,
via
https://github.com/honeycombio/infra/blob/main/terraform/modules/aws_integrations/rds_events.tf),
see
https://ui.kibble-eu1.honeycomb.io/honeycomb/environments/dogfood-eu1/datasets/rds-events/result/kz-S1g9i2LGo?tab=events
(data generated by manually creating a db snapshot). For the trigger we
intend to use this with, see
https://github.com/honeycombio/infra/pull/5552 .