maint: Update publish regions list and make it more maintainable (#262)
## Short description of the changes
The regions list is badly out of date, and that's causing issues for a
customer.
This PR updates the list based on regions that our AWS account is opted
in to, and also ones where the S3 bucket exists. It calls out regions
where the S3 bucket does not exist and provides instructions.

## How to verify that this has the expected result
Pretty difficult to know if it'll work or not till we merge. I did run
shellcheck on the script and it raised other concerns, but nothing about
the parts I changed.

---------

Signed-off-by: Irving Popovetsky <irving@honeycomb.io>